Some qBound vs. qreal fixes
REVIEW: 125996

+++ examples/cpp/marbleQuick2/main.cpp
@@ -58,7 +58,7 @@ private:
     void makePinch(QPointF center, Qt::GestureState state, qreal scale = 1)
     {
         scale = sqrt(sqrt(scale));
-        scale = qBound(0.5, scale, 2.0);
+        scale = qBound(static_cast<qreal>(0.5), scale, static_cast<qreal>(2.0));
         pinch(center, scale, state);
     }
 };
